+++ c7n/actions/__init__.py
@@ -9,5 +9,5 @@ from .metric import PutMetric
 from .network import ModifyVpcSecurityGroupsAction
 from .notify import BaseNotify, Notify
 from .policy import RemovePolicyBase, ModifyPolicyBase
-from .resize import ResizeAutoscalingTarget
+from .autoscaling import AutoscalingBase
